Add files via upload
Refactored notification options to prompt user and save all settings to a configuration file rather than relying on the user to edit the original script and saving some settings to two files.  File also adds a separate scriplet to update the definitions, and notifies user at runtime if there is an updated version of this script available.